Спринт 16. Начало. Создал новые пакеты, старые удалил, предварительно все из них вытащив
