+++ app/src/main/java/com/example/playlistmaker/Creator.kt
@@ -1,18 +1,18 @@
 package com.example.playlistmaker
 
-import com.example.playlistmaker.data.network.RetrofitNetworkClient
-import com.example.playlistmaker.data.repository.TrackMapper
-import com.example.playlistmaker.data.repository.TracksRepositoryImpl
-import com.example.playlistmaker.domain.api.TracksInteractor
-import com.example.playlistmaker.domain.api.TracksRepository
-import com.example.playlistmaker.domain.impl.TracksInteractorImpl
+import com.example.playlistmaker.data.RetrofitNetworkClient
+import com.example.playlistmaker.data.TrackMapper
+import com.example.playlistmaker.data.TracksRepositoryImpl
+import com.example.playlistmaker.domain.TracksInteractor
+import com.example.playlistmaker.domain.TracksRepository
+import com.example.playlistmaker.domain.TracksInteractorImpl
 
 object Creator {
-    private fun getTracksRepository(): TracksRepository{
+    private fun getTracksRepository(): TracksRepository {
         return TracksRepositoryImpl(RetrofitNetworkClient(), trackMapper = TrackMapper())
     }
 
-    fun provideTracksInteractor(): TracksInteractor{
+    fun provideTracksInteractor(): TracksInteractor {
         return TracksInteractorImpl(getTracksRepository())
     }
 }
