+++ app/src/main/java/com/example/playlistmaker/TrackView.kt
@@ -1,0 +1,4 @@
+package com.example.playlistmaker
+
+interface TrackView {
+}
