+++ app/src/main/java/com/example/playlistmaker/data/ITunesAPI.kt
@@ -1,6 +1,5 @@
-package com.example.playlistmaker.data.network
+package com.example.playlistmaker.data
 
-import com.example.playlistmaker.data.dto.TrackSearchResponse
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
