+++ app/src/main/java/com/example/playlistmaker/data/NetworkClient.kt
@@ -1,7 +1,5 @@
 package com.example.playlistmaker.data
 
-import com.example.playlistmaker.data.dto.Response
-
 interface NetworkClient {
     fun doRequest(dto: Any): Response
 }
