+++ app/src/main/java/com/example/playlistmaker/data/Responce.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data.dto
+package com.example.playlistmaker.data
 
 open class Response() {
     var resultCode = 0
