+++ app/src/main/java/com/example/playlistmaker/data/TrackDto.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data.dto
+package com.example.playlistmaker.data
 
 data class TrackDto(
     val trackId: String,
