+++ app/src/main/java/com/example/playlistmaker/data/TrackSearchResponse.kt
@@ -1,3 +1,3 @@
-package com.example.playlistmaker.data.dto
+package com.example.playlistmaker.data
 
 class TrackSearchResponse(val results: List<TrackDto>) : Response()
