+++ app/src/main/java/com/example/playlistmaker/data/TracksRepositoryImpl.kt
@@ -1,0 +1,19 @@
+package com.example.playlistmaker.data
+
+import com.example.playlistmaker.domain.TracksRepository
+import com.example.playlistmaker.domain.Track
+
+class TracksRepositoryImpl(private val networkClient: NetworkClient, private val trackMapper: TrackMapper):
+    TracksRepository {
+
+    override fun searchTrack(expression: String): List<Track> {
+        val response = networkClient.doRequest(TrackSearchRequest(expression))
+        if (response.resultCode == 200){
+            return (response as TrackSearchResponse).results.map { trackDto ->
+                trackMapper.mapToDomain(trackDto)
+            }
+        } else {
+            return emptyList()
+        }
+    }
+}
