--- app/src/main/java/com/example/playlistmaker/data/repository/TracksRepositoryImpl.kt
@@ -1,21 +1,0 @@
-package com.example.playlistmaker.data.repository
-
-import com.example.playlistmaker.data.NetworkClient
-import com.example.playlistmaker.data.dto.TrackSearchRequest
-import com.example.playlistmaker.data.dto.TrackSearchResponse
-import com.example.playlistmaker.domain.api.TracksRepository
-import com.example.playlistmaker.domain.models.Track
-
-class TracksRepositoryImpl(private val networkClient: NetworkClient, private val trackMapper: TrackMapper): TracksRepository {
-
-    override fun searchTrack(expression: String): List<Track> {
-        val response = networkClient.doRequest(TrackSearchRequest(expression))
-        if (response.resultCode == 200){
-            return (response as TrackSearchResponse).results.map { trackDto ->
-                trackMapper.mapToDomain(trackDto)
-            }
-        } else {
-            return emptyList()
-        }
-    }
-}
