+++ app/src/main/java/com/example/playlistmaker/domain/Track.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.domain.models
+package com.example.playlistmaker.domain
 
 import java.text.SimpleDateFormat
 import java.util.Locale
