+++ app/src/main/java/com/example/playlistmaker/domain/TracksInteractor.kt
@@ -1,6 +1,4 @@
-package com.example.playlistmaker.domain.api
-
-import com.example.playlistmaker.domain.models.Track
+package com.example.playlistmaker.domain
 
 interface TracksInteractor {
     fun searchTrack(expression: String, consumer: TrackConsumer)
