--- app/src/main/java/com/example/playlistmaker/presentations/tracks/TrackView.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker.presentations.tracks
-
-interface TrackView {
-}
