+++ app/src/main/java/com/example/playlistmaker/ui/MainActivity.kt
@@ -1,14 +1,11 @@
-package com.example.playlistmaker.ui.main
+package com.example.playlistmaker.ui
 
 import android.annotation.SuppressLint
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.Button
-import com.example.playlistmaker.ui.media.MediaActivity
 import com.example.playlistmaker.R
-import com.example.playlistmaker.ui.settings.SettingsActivity
-import com.example.playlistmaker.ui.tracks.SearchActivity
 
 class MainActivity : AppCompatActivity() {
