+++ app/src/main/java/com/example/playlistmaker/ui/MediaActivity.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.ui.media
+package com.example.playlistmaker.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
