+++ app/src/main/java/com/example/playlistmaker/ui/TrackAdapter.kt
@@ -1,10 +1,10 @@
-package com.example.playlistmaker.ui.tracks
+package com.example.playlistmaker.ui
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.playlistmaker.R
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.Track
 
 class TrackAdapter(private var tracks: List<Track>, val listener: RecycleViewListener) : RecyclerView.Adapter<TrackViewHolder>() {
